Mesh Rendering & Material Support

+++ Tupla-Engine/Source/Tupla/AssetManager/AssetManager.cpp
@@ -9,5 +9,5 @@ void Tupla::AssetManager::InsertFilePath(const std::string& phyiscalFile, const 
 void Tupla::AssetManager::ReadBinaryFromFile(const std::string& virtualFile, std::vector<char>& buffer)
 {
 	auto mapping = m_VirtualToPhysicalMap[HASH_RUNTIME_STR(virtualFile)];
-	CommonUtilities::ReadFileBinary(mapping.PhysicalFilePath.c_str(), buffer, mapping.Offset, mapping.Size);
+	CommonUtilities::ReadFileBinary(mapping.PhysicalFilePath.c_str(), buffer, (u32)mapping.Offset, (u32)mapping.Size);
 }
